requirements.txt
Babel==2.4.0
Django==1.9.6
Pillow==3.0.0
astroid==1.4.9
beautifulsoup4==4.5.1
colorama==0.3.7
coverage==4.3.4
dj-database-url==0.4.2
dj-static==0.0.6
django-address==0.1.5
django-log-request-id==1.3.2
django-phonenumber-field==1.3.0
django-storages==1.5.2
flake8==2.6.2
gunicorn==19.7.1
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.5.3
pep8==1.7.0
phonenumbers==8.4.1
psycopg2==2.6.1
pycodestyle==2.0.0
pyflakes==1.2.3
pylint==1.6.5
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pyparsing==2.2.0
python-decouple==3.0
pytz==2017.2
requests==2.13.0
six==1.10.0
static3==0.7.0
whitenoise==3.3.0
wrapt==1.10.10
